# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelZone Backend",
    "description": "Unified SOC backend: Phase 24 (API), 26 (Incidents), 27 (Notifications)",
    "version": "0.28.0"
  },
  "paths": {
    "/v1/incidents/{incident_id}/case-pack": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Export Case Pack",
        "operationId": "export_case_pack_v1_incidents__incident_id__case_pack_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/event-sets": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Event Set",
        "operationId": "event_set_v1_event_sets_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventSetIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/replay/rules": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Rules",
        "operationId": "rules_v1_replay_rules_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/replay": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Run",
        "operationId": "run_v1_replay_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/replay/{run_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Run",
        "operationId": "get_run_v1_replay__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assets/{host_id}/snapshots": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Snapshot",
        "operationId": "snapshot_v1_assets__host_id__snapshots_post",
        "parameters": [
          {
            "name": "host_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Host Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetStateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assets/{host_id}/what-changed": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Changes",
        "operationId": "changes_v1_assets__host_id__what_changed_get",
        "parameters": [
          {
            "name": "host_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Host Id"
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/incidents/{incident_id}/merge": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Merge Incident",
        "description": "Operator: merges now. Analyst: files a PENDING merge request that an operator must approve.",
        "operationId": "merge_incident_v1_incidents__incident_id__merge_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/merge-requests": {
      "get": {
        "tags": [
          "merge-requests"
        ],
        "summary": "List Requests",
        "operationId": "list_requests_v1_merge_requests_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/merge-requests/{mid}/approve": {
      "post": {
        "tags": [
          "merge-requests"
        ],
        "summary": "Approve",
        "operationId": "approve_v1_merge_requests__mid__approve_post",
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/merge-requests/{mid}/reject": {
      "post": {
        "tags": [
          "merge-requests"
        ],
        "summary": "Reject",
        "operationId": "reject_v1_merge_requests__mid__reject_post",
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetStateIn": {
        "properties": {
          "local_users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Local Users"
          },
          "services": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Services"
          },
          "startup_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Startup Items"
          },
          "unsigned_processes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Unsigned Processes"
          },
          "outbound_destinations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Outbound Destinations"
          },
          "telemetry_sources": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "healthy",
                "degraded",
                "offline",
                "stopped",
                "unknown"
              ]
            },
            "type": "object",
            "title": "Telemetry Sources"
          },
          "criticality": {
            "type": "string",
            "enum": [
              "low",
              "normal",
              "high",
              "critical"
            ],
            "title": "Criticality",
            "default": "normal"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AssetStateIn"
      },
      "EventSetIn": {
        "properties": {
          "event_uids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10000,
            "minItems": 1,
            "title": "Event Uids"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "REAL",
              "TEST"
            ],
            "title": "Execution Mode",
            "default": "REAL"
          }
        },
        "type": "object",
        "required": [
          "event_uids"
        ],
        "title": "EventSetIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MergeIn": {
        "properties": {
          "target_id": {
            "type": "string",
            "title": "Target Id"
          }
        },
        "type": "object",
        "required": [
          "target_id"
        ],
        "title": "MergeIn"
      },
...
      "ReplayIn": {
        "properties": {
          "event_set_id": {
            "type": "string",
            "title": "Event Set Id"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id",
            "default": "asset-correlation"
          },
          "old_version": {
            "type": "string",
            "title": "Old Version",
            "default": "1"
          },
          "new_version": {
            "type": "string",
            "title": "New Version",
            "default": "2"
          }
        },
        "type": "object",
        "required": [
          "event_set_id"
        ],
        "title": "ReplayIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}